pyproject.toml
[tool.poetry]
name = "searchadscli"
version = "0.1.0"
description = "A CLI for running Apple Search Ads Advanced campaigns."
authors = ["rkotzy <yolo@ryan.wtf>"]
readme = "README.md"

[tool.poetry.scripts]
searchadscli = "searchadscli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
authlib = "^1.2.1"
pycryptodome = "^3.18.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
